backend/requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.1.1
et-xmlfile==1.1.0
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.1
isort==5.9.2
itypes==1.2.0
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.8.0
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
passlib==1.7.4
Pillow==8.3.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==4.4.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
recipes==0.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
slugify==0.0.1
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
